requirement.txt
faster-whisper
markdown2
pdfkit
llama-cpp-python
